Refactor to expose orchestrator client
