+++ examples/infra/mod.rs
@@ -6,7 +6,6 @@ use async_lock::RwLock;
 use async_trait::async_trait;
 use clap::Parser;
 use futures::Future;
-use futures::FutureExt;
 use hotshot::{
     traits::{
         implementations::{
@@ -19,6 +18,7 @@ use hotshot::{
 };
 use hotshot_orchestrator::{
     self,
+    client::{OrchestratorClient, ValidatorArgs},
     config::{NetworkConfig, NetworkConfigFile, WebServerConfig},
 };
 use hotshot_types::traits::election::ConsensusExchange;
@@ -58,10 +58,8 @@ use std::{
     num::NonZeroUsize,
     str::FromStr,
     sync::Arc,
-    time::{Duration, Instant},
+    time::Instant,
 };
-use surf_disco::error::ClientError;
-use surf_disco::Client;
 #[allow(deprecated)]
 use tracing::error;
 
@@ -152,24 +150,6 @@ pub async fn run_orchestrator<
         TYPES::ElectionConfigType,
     >(run_config, host, port)
     .await;
-}
-
-// VALIDATOR
-
-#[derive(Parser, Debug, Clone)]
-#[command(
-    name = "Multi-machine consensus",
-    about = "Simulates consensus among multiple machines"
-)]
-/// Arguments passed to the validator
-pub struct ValidatorArgs {
-    /// The address the orchestrator runs on
-    pub host: IpAddr,
-    /// The port the orchestrator runs on
-    pub port: u16,
-    /// This node's public IP address, for libp2p
-    /// If no IP address is passed in, it will default to 127.0.0.1
-    pub public_ip: Option<IpAddr>,
 }
 
 /// Defines the behavior of a "run" of the network with a given configuration
@@ -760,106 +740,6 @@ where
 
     fn get_config(&self) -> NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType> {
         self.config.clone()
-    }
-}
-
-/// Holds the client connection to the orchestrator
-pub struct OrchestratorClient {
-    client: surf_disco::Client<ClientError>,
-}
-
-impl OrchestratorClient {
-    /// Creates the client that connects to the orchestrator
-    pub async fn connect_to_orchestrator(args: ValidatorArgs) -> Self {
-        let base_url = format!("{0}:{1}", args.host, args.port);
-        let base_url = format!("http://{base_url}").parse().unwrap();
-        let client = surf_disco::Client::<ClientError>::new(base_url);
-        assert!(client.connect(None).await);
-
-        OrchestratorClient { client }
-    }
-
-    /// Sends an identify message to the server
-    /// Returns this validator's node_index in the network
-    pub async fn identify_with_orchestrator(&self, identity: String) -> u16 {
-        let identity = identity.as_str();
-        let f = |client: Client<ClientError>| {
-            async move {
-                let node_index: Result<u16, ClientError> = client
-                    .post(&format!("api/identity/{identity}"))
-                    .send()
-                    .await;
-                node_index
-            }
-            .boxed()
-        };
-        self.wait_for_fn_from_orchestrator(f).await
-    }
-
-    /// Returns the run configuration from the orchestrator
-    /// Will block until the configuration is returned
-    pub async fn get_config_from_orchestrator<TYPES: NodeType>(
-        &self,
-        node_index: u16,
-    ) -> NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType> {
-        let f = |client: Client<ClientError>| {
-            async move {
-                let config: Result<
-                    NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>,
-                    ClientError,
-                > = client
-                    .post(&format!("api/config/{node_index}"))
-                    .send()
-                    .await;
-                config
-            }
-            .boxed()
-        };
-        self.wait_for_fn_from_orchestrator(f).await
-    }
-
-    /// Tells the orchestrator this validator is ready to start
-    /// Blocks until the orchestrator indicates all nodes are ready to start
-    pub async fn wait_for_all_nodes_ready(&self, node_index: u64) -> bool {
-        let send_ready_f = |client: Client<ClientError>| {
-            async move {
-                let result: Result<_, ClientError> = client
-                    .post("api/ready")
-                    .body_json(&node_index)
-                    .unwrap()
-                    .send()
-                    .await;
-                result
-            }
-            .boxed()
-        };
-        self.wait_for_fn_from_orchestrator::<_, _, ()>(send_ready_f)
-            .await;
-
-        let wait_for_all_nodes_ready_f = |client: Client<ClientError>| {
-            async move { client.get("api/start").send().await }.boxed()
-        };
-        self.wait_for_fn_from_orchestrator(wait_for_all_nodes_ready_f)
-            .await
-    }
-
-    /// Generic function that waits for the orchestrator to return a non-error
-    /// Returns whatever type the given function returns
-    pub async fn wait_for_fn_from_orchestrator<F, Fut, GEN>(&self, f: F) -> GEN
-    where
-        F: Fn(Client<ClientError>) -> Fut,
-        Fut: Future<Output = Result<GEN, ClientError>>,
-    {
-        loop {
-            let client = self.client.clone();
-            let res = f(client).await;
-            match res {
-                Ok(x) => break x,
-                Err(_x) => {
-                    async_sleep(Duration::from_millis(250)).await;
-                }
-            }
-        }
     }
 }
