+++ examples/web-server-da/validator.rs
@@ -1,11 +1,11 @@
-use async_compatibility_layer::logging::{setup_logging, setup_backtrace};
+use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use clap::Parser;
 use hotshot::demos::sdemo::SDemoTypes;
 use tracing::instrument;
 
 use crate::types::{DANetwork, NodeImpl, QuorumNetwork, ThisMembership, ThisRun};
 
-use crate::infra::ValidatorArgs;
+use hotshot_orchestrator::client::ValidatorArgs;
 
 pub mod types;
